+++ NEWS.md
@@ -1,4 +1,4 @@
-# jlme (development version)
+# jlme 0.4.0
 
 * New argument `jlmer(fit = TRUE)`. If `FALSE`, returns the unfit model object.